fix pd-sep: request key, cache key, error code, net util, tp size

+++ maga_transformer/cpp/model_rpc/DecodeRpcServer.h
@@ -25,7 +25,7 @@ public:
     class LoadKVCacheContext {
     public:
         LoadKVCacheContext(int64_t request_id, const std::string& request_key, const std::string& peer_ip,
-                            const std::vector<int32_t>& cache_keys, const std::vector<int32_t>& block_ids,
+                            const std::vector<int64_t>& cache_keys, const std::vector<int32_t>& block_ids,
                             int64_t reuse_block_size, int64_t timeout_ms, grpc::ServerContext* server_context) :
                             request_id(request_id), request_key(request_key), peer_ip(peer_ip),
                             cache_keys(cache_keys), block_ids(block_ids), reuse_block_size(reuse_block_size),
@@ -33,7 +33,7 @@ public:
         int64_t request_id;
         const std::string& request_key;
         const std::string& peer_ip;
-        const std::vector<int32_t>& cache_keys;
+        const std::vector<int64_t>& cache_keys;
         const std::vector<int32_t>& block_ids;
         int64_t reuse_block_size;
         int64_t timeout_ms;
